Added additional views to break up the AccountView
